Fix issue where analyzers are missing when computing diagnostics OOP
If you were to have the same analyzer added as both a host analyzer
and a project analyzer, we might end up removing it from the "project
analyzer" set when we create the project CompilationWithAnalyzers. This
confuses later code since it expects the analyzers to be present.

This was regressed in c860af4093a89d0d09bc6a11baca592a013abb3c,
where a mechanical refactoring took us from having a single
CompilationWithAnalyzers to two, one for project analyzers and one for
host analyzers. The loop in the method was copied into two, but the
set wasn't cleared between the two loops. I believe the intent was to
clear out the set (rather than the set persist to deal with duplicates
between the host and project sets), given the comment implies we do
want duplicates to overwrite in the analyzerMap that we're also
producing.

I also split the two analyzer maps, one for project analyzers and one for
host analyzers. If you don't do this, it means that a conflict where the
same analyzer existing in both sets will overwrite each other, later
code will then return a project analyzer which might be a different
isntance than the host analyzer equivalent.