Reduce allocations in DiagnosticDescriptor.GetEffectiveSeverity
This method shows up as about 0.9% of allocations in the completions scenarios in the razor CompletionInCohosting speedometer test.

Instead, as called out in an *old* performance issue that was closed long ago, the CompilationOptions.Filter method could be forked such that the new version didn't need to take in a dummy diagnostic and instead worked off the DiagnosticDescriptor.

Addresses https://github.com/dotnet/roslyn/issues/3650