Reduce allocations obtaining classified spans in ClassifierHelper
ClassifiedSpan[] shows up as 0.7% of total allocations (in LOH) in the completion scenarios in the razor cohosting speedometer test.

Surrounding code already uses a custom pool for ClassifiedSpan[] via Classifier.GetPooledList. This PR changes ClassifierHelper.GetClassifiedSpansAsync to take in a SegmentedList<ClassifiedSpan> to populate rather than returning an ImmutableArray<ClassifiedSpan>.