Redo how and when we report source generator telemetry
Previously, we were triggering a report of source generator telemetry
when the user closed the solution. The intent of this was to be
able to do it prior to "context" that the telemetry service has set
is cleared; this means we could understand the performance for one
solution versus another.

This approach however was broken once we moved the computation of source
generators OOP; this meant we'd report the 'in-process' results,
which are almost always empty.

This switches to reporting the results for all out-of-process and in-
process runs. Unfortunately, the trick of us doing it during solution
close isn't possible without blocking the UI thread synchronously on
OOP which we don't want to do, so for now we'll just report telemetry
every few minutes as long as there's new telemetry to report. There is
work happening that would allow us to have an asynchronous handler
there again which we can look at moving back to later.