Add Portable PDB No References Option
The `/debugportablenoreferences` command-line option for `csc`/`vbc` (or `DebugPortableNoReferences` in the `Csc` or `Vbc` MSBuild task) can be used to not emit this section.

Consider when an assembly changes in a way that does not affect some/all of the consumers.
For example if a new method/class/etc is added (if using ref assemblies this only applies to public entites).

When a consumer re-compiles all of its debugging information and IL will come out identically (assuming deterministic compilation is used).
However, this section will change content since the MVID in of one of its dependencies changed.
This causes the PDB to change GUIDs which in turn causes the assembly to change PDB Guids and MVID.
All of these new DLLs/PDBs need to be copied, indexed, etc which slows down build time and increase space usage.
With this section disabled, the consumers may produce the same content (assuming the change to the dependency
doesn't affect them) and will reduce how much further build steps need to happen.

However, the PDB no longer has the information about the exact MVID/details of each of the dependencies making it not possible to reconstruct the exact inputs to compilation from just the PDB.
If you need this functionality you should not use this flag and you by-design want new PDBs any time a dependency changes.

This option is motivated by users of roslyn who want to use portable PDBs but do not want to pay for it with slower builds and do not use the reconstruction data.