Ensure that CheckValue isn't called twice on lambda query receivers
When creating query expressions, most of the time the receiver the query is not checked for being a valid RValue by calling code. However, in one case, it is; this can then cause an assert that we are checking for field-backed property access multiple times. Fixes https://github.com/dotnet/roslyn/issues/80008.