Don't cache known-broken compositions
Right now we have an issue where users are getting broken MEF
compositions cached on their machines, so subsequent launches of the
product are operating on a broken composition. It's not "broken"
in the sense that "the cache cannot be loaded" but rather the content
of the cache is just missing most of our MEF parts. As a defense-in-
depth change to deal with potential transient issues, let's not cache
compositions we know are bad.