Switch to using IVsAsyncSolutionEventListener for batch scope creation
For the legacy project systems, we implicitly create a batch scope
around each created project so that way all the changes to a project
don't create an excess of workspace changes. We complete these batches
in OnAfterSolutionLoad. VS now has an asynchronous event which we can
respond to instead which allows for us to complete the batches
asynchronously. Although this is still considered part of the solution
load process and shouldn't change the total time, it'll at least ensure
the UI thread can stay responsive during the load.

To simplify testing (and since there's now an easy way to ask if a
solution is currently in the middle of opening) I'm switching the
_solutionLoaded flag to _solutionLoading so the lifetime we're creating
the batches in is clearer.