Look at all composition errors, not just the first set of them
The CompositionErrors collection is ordered -- the first set of errors
are the "root" errors, then there are follow up sets that are downstream
failures from those first ones. Since our code was only looking at the
first set, it's possible that an "expected" failure there might still
cause downstream issues which could break the rest of the system.