Avoid loading MEF caches that are clearly bad
Right now we're seeing an issue where sometimes we have invalid
cache files written that don't contain the parts we expect them
to have. This causes us to be unable to create our ServiceHub's
services. As a tactical workaround, the files that are bad tend
to be just a few kilobytes in size, but the ones that are valid
are much bigger. This allows us to easily filter out bad ones.