Fix our registration for our TempPE compilers
Commit 1b4f5d798e0c9b9df3e305c0310590224a8f323d removed the service
registration from our pkgefs, but this was incorrect. We don't run
the CreatePkgDef tool, so the attributes defined on the package
don't actually register the service into the system. The refactoring
in 391e50f027c238a619a4a6ec3fdf04b1eed8bce4 also lost the promote: true
arguments, so the service registration didn't work either.

Since we didn't have any tests on this, this adds some tests directly
testing the TempPE services so we don't break this again.