Introduce IProjectFileInfoLoaderFactory abstraction
Introduce an IProjectFileInfoLoaderFactory abstraction that can be optionally passed to an MSBuildProjectLoader to allow customization of ProjectFileInfo loading. If an IProjectFileInfoLoaderFactory isn't provided, the default BuildHostProjectFileInfoLoaderFactory instance is used.

This abstraction is intended to allow dotnet-watch to provide a custom IProjectFileLoaderFactory that can produce ProjectFileInfos without using an out-of-proc build host.