Add the ability to preload parts of the project system
For the csproj and msvbprj project systems, our COM services get called
on the UI thread when project information is being sent our way. This
has the side effect of creating all our "lazily created" MEF services
on that UI thread during that time, which can result in more unnecessary
blocking. Changing how those project systems work would be a hugely
complicated change, but what we can do is take advantage of the
preloading support that exists in VS to ensure we preload what we can
off the UI thread before we hit the legacy code.