Fix add missing using from throwing exceptions in serialization
When we serialize types to our ServiceHub process, we use several
ForceTypelessFormatters, which specifies that we want to support
polymorphism when serializing/deserializing. To deserialize the type on
the ServiceHub side, MessagePack calls Type.GetType() passing in the
name of the type to be deserialized. Such a call, on .NET Core,
must happen with the context of some AssemblyLoadContext, but today
nothing in StreamJsonRpc or MessagePack specifies which
AssemblyLoadContext that should be. So, if we get unlucky, it'll try
it in the default AssemblyLoadContext which will fail.

The workaround/fix here is to call EnterContextualReflection which
lets us specify that when we are deserializing these types, we want
the deserialization to happen within the AssemblyLoadContext that
holds Roslyn's service, as opposed the default AssemblyLoadContext in
ServiceHub.

We investigated whether this fix should be somewhere else than Roslyn,
since Roslyn isn't doing anything wrong here -- anybody else using
StreamJsonRpc and MessagePack inside of an AssemblyLoadContext could
run into the same issue. Since this is a fairly impacting issue, we're
going with a fix here, since any deeper change would require fixes in
either of those libraries which would then have to be inserted
into other repositories in order to unblock us. And even then, those
fixes might still require us to specify the AssemblyLoadContext we want
to be used for deserialization, so we'd still have to make a Roslyn
change anyways. There's no clear benefit for all of that work when
we can be direct with what we want here, so this is the easiest to do.

The reason we are running into this now is unclear -- our assumption
is a recent change to ServiceHub means either how our
AssemblyLoadContext is being created changed, or some common dependency
in our AssemblyLoadContext and the default AssemblyLoadContext changed.
If we had some shared dependency but same/different versions, it's
possible that there was some mixing of dependencies across the
AssemblyLoadContext which uncovered this issue. The fact we're not
being explicit about the AssemblyLoadContext though is really not
something ServiceHub can address (since we're configuring the RPC
channel), so understanding the triggering change isn't terribly
important since the fix is the same no matter what.

Fixes https://devdiv.visualstudio.com/DevDiv/_workitems/edit/2576636