Update NgenPriority to reduce number of priority 1 assemblies  (#80411)
Changed two assemblies to be priority 2 to lower the duration of ngen for priority 1 assemblies. We are trying to improve how we ngen assemblies during installation to improve VS's startup, and as part of this are reducing the number of assemblies that have a priority of 1. JIT time during VS startup was used to determine the priorities, the following two assemblies are being changed to be priority 2:
- msbuild\current\bin\roslyn\microsoft.codeanalysis.csharp.dll
- msbuild\current\bin\roslyn\microsoft.codeanalysis.dll