Reduce allocations in SourceText.ToString
This method was using a PooledStringBuilder to do it's string buildup. When a PooledStringBuilder's underlying StringBuilder exceeeds 1024, it is no longer returned to the pool upon completion. SourceText.ToString results commonly exceeed 1024 characters. To better improve those cases, we can use string.Create directly to avoid the StringBuilder altogether.

This accounts for about 0.6% of all allocations in the RoslynCodeAnalysisService process in the razor CompletionInCohosting speedometer test.