Deprioritize the .editorconfig designer (but let you still pick it)
This reduces the .editorconfig designer's default priority so users
no longer get it as a default if they double click an .editorconfig
file. They can still choose "open with" to open it explicitly if needed.
The goal being this is a soft deprecation, and we can go ahead with
complete removal if we see usage is sufficiently small.

To allow a user to actually pick the other option, I had to add a string
specifying the name of the designer.