Update the condition used to include the CodeStyle .globalconfig
The logic of the original check seems to be a bit confused. It seems reasonable to include the globalconfig if the build has specifically called for it by setting either AnalysisLevelStyle or AnalysisLevelMode prior to our target running. In addition we keep the AnalysisLevel gate that will automatically add the configuration when the level is high enough.