When searching for the original file from a PDB, only allow absolute paths
Without this check, if the PDB in question used path map to create a
relative path, it's possible for this check to pass and end up
handing around relative paths to the rest of the system (and the VS
Shell). This eventually causes other things to break that expect full
paths.

Given a relative path would only work here by "luck" if the devenv
current working directory happened to line up, it seems easiest just
to assert it's an absolute path before going further.

Fixes https://github.com/dotnet/roslyn/issues/79014