Try to reuse initialized generator drivers when we can
We're seeing a performnace issue in Razor projects in cohosting
scenarios. The first run of the Razor generator is relatively slow,
and so ideally we wouldn't have that happen more than once. However,
since multiple features might request compilations for a Razor project
during startup in parallel, it's possible that all of those parallel
invocations might initialize the generator multiple times if they
happen to be different Solution instances. This change tries to ensure
that when we're creating a GeneratorDriver for a project when we don't
already have one, we'll do that computation just once, and try to share
that across the forks. This ensures we do the expensive initialization
only once.