Extensions: Add ReduceExtensionMember API (#80659)
Closes https://github.com/dotnet/roslyn/issues/80273

There is a small change in behavior when dealing with extensions that
aren't fully inferred. Previously, `GetCompatibleSubstitutedMember`
dropped them. Now, it's up to the caller to decide.
The `LookupSymbols_WasNotFullyInferred` test illustrates such a change.
For lookup, we do want to return members that can't be fully inferred.

Relates to test plan https://github.com/dotnet/roslyn/issues/76130