Fix go to definition for conversions on invalid constructor overloads (#78514)
Closes #73498
Closes #77545

Fixes one part of #75833, the appropriate tests were adjusted to reflect
the new behavior

With this PR, the `GetSymbolInfo` API and all the corresponding APIs
using it will be affected for the scenarios where an object creation
expression used a constructor overload that doesn't exist, whilst the
expression was being implicitly converted to another type. The previous
behavior would resolve to the target type of the conversion, since the
binding result would be an overload resolution, whereas now it has been
changed to resolve to the existing overload resolution result.

The goal of this change is also reflected in GTD which will now
correctly navigate to the type whose constructor was not found, instead
of the target conversion type, or the implicit conversion method.

This also adds tests for explicit conversions on object creation
expressions, covering the above scenaria, as well as testing the
existing correct behaviors of when there is no conversion, or when there
is no overload resolution failure.

---------

Co-authored-by: AlekseyTs <AlekseyTs@users.noreply.github.com>