Make string concat assert more precise (#80619)
These asserts are intended to ensure that, if we ever come across a scenario that should be optimized, we consider that. The current assert is broader than this though, and already isn't in line with the code it reflects. This updates the assert to be clearer. Fixes https://github.com/dotnet/roslyn/issues/80254.