Fix issue with when a CR/LF split occurs at a SubText's underlying span start (#80911)
I hit an exception in the SourceText system when inserting a copilot
edit for C# content in a razor file when the file uses newlines only.
This ended up causing a CR/LF split in the subtext class, which had some
bogus logic for when that happened.

Added a very simple unit test that demonstrated previous failure.