Add support for lineFoldingOnly folding ranges (#80955)
Fixes https://github.com/dotnet/vscode-csharp/issues/7974
Unblocks
https://github.com/dotnet/roslyn/pull/80944#issuecomment-3459077743

VSCode sets the
[lineFoldingOnly](https://microsoft.github.io/language-server-protocol/specifications/lsp/3.17/specification/#foldingRangeClientCapabilities)
LSP capability, indicating it cannot handle multiple folding ranges that
start on the same line as another range starts or ends on. Folding
ranges that are not compatible are simply dropped.

While we cannot always do anything more than drop one of the ranges, we
can choose which range to drop and we can sometimes adjust ranges so
that they will not be dropped:
1. If two ranges start on the same line, we want to pick the inner range
for folding.
2. If a range ends on the same line as another starts, we can attempt to
shrink it by one line to handle common cases like an else block starting
on the same line as the if block ended:
```
if (blah) {
    ...
} else {
    ...
}
```

This is done as a post processing step of the ranges. This seemed
reasonable as the service generally independently constructs folding
ranges based on the syntax and has little knowledge of what the other
ranges will be produced (to know if they overlap)


![folding_ranges_overlap](https://github.com/user-attachments/assets/8c712443-faea-4d9d-a59d-1bb456860740)