#81118: Fix dropdown in "Move static members" dialog (#81119)
Fixes #81118.

Apparently the dialog broke when the `TypeNameItem` type was moved to
another namespace. This fixes that reference and fixes up some of the
bindings that are used in the dialog (and also sets `Mode=OneTime` since
the bound properties are readonly and so the model does not change while
the dialog is open).

<img width="486" height="491" alt="grafik"
src="https://github.com/user-attachments/assets/c7de8b8a-3a4b-43cf-aa27-93541f84e35f"
/>