Simplify sequence point display (#81104)
Right now, two different string parameters are used to control sequence
point display. Neither string is used beyond null checks. I've
simplified this to be just a single, more explanatory enum. Along the
way, I noticed that BasicBlock's visualization is also broken, so I
restored the private method that it used for that.