Include default RSP files for core compilers (#80993)
Follow up on https://github.com/dotnet/roslyn/pull/79911.

Fixes another inconsistency between the core and fx compilers: if the
`<Csc>` task is used manually on a C# file which has `using
System.Linq;` (for example), the fx compiler handles that fine (because
there is `csc.rsp` deployed with it which has a set of default
references) whereas the core compiler fails. A real-world use-case can
be seen in
https://dev.azure.com/devdiv/DevDiv/_git/VS/pullrequest/684436.

The fix is to include this default `csc.rsp` in scenarios where we
invoke the core compiler on fx msbuild (to avoid a break there). We
continue to not include that `csc.rsp` in scenarios where just the core
csc is being used (again, to avoid a break in those scenarios).