Document excerpt for Razor, and fix Find All Refs (#81291)
Part of https://devdiv.visualstudio.com/DevDiv/_workitems/edit/2633418

We had to revert some code that made FAR, Hover, Go To Def etc. work in
Razor files because to support it we had to add `#line` mappings for
things, and that broke Hot Reload by changing PDB offsets. Who knrew!

This PR enables us to bring back that support without `#line` directives
by allowing locations in Razor source genned documents to be found by
FAR even if they're hidden. It also tweaks the LSP source mapping code
to drop results that can't be mapped, which actually matches was the VS
side already did anyway. Finally it exposes a hook for Razor to
implement a document excerpt service, which is something I punted
previously because the results with the `#line` directives were good
enough, but now they're really not.

Commit-at-a-time review is probably simpler, to make it clear which
change is for which purpose.