Improve LSP conformance in MS.CA.LanguageServer (#81174)
I am proposing fixes for two LSP conformance issues I discovered while
integrating Roslyn LSP with Zed:

1. The response from the `textDocument/diagnostic` request cannot be
`null`. MS.CA.LanguageServer currently returns a `null` response in the
case described here:
https://github.com/dotnet/roslyn/blob/76a39c1a97fe5c73549786fc6944dc188a7b7b3b/src/LanguageServer/Protocol/Handler/Diagnostics/DiagnosticSourceProviders/DiagnosticSourceManager.cs#L72-L80
Because the set of diagnostic handlers is empty, a response value is
never generated for the request, and we fall through to returning
`null`. Instead, explicitly return an empty list of diagnostics. (This
is exactly how VS Code's LSP client treats a `null` response.)

2. Methods that take no parameters (e.g. `workspace/diagnostic/refresh`)
were being sent with an empty array as the parameter value. This is
because the `InvokeWithCancellationAsync` method in StreamJsonRpc
normalizes unspecified arguments to an empty list. Use the
`InvokeParameterObjectAsync` method instead, which don't do this.