Report basic text changes if span mapping isn't available (#81492)
I stuffed up, in https://github.com/dotnet/roslyn/pull/81450 I map LSP
rename map edits so that Razor components could be renamed, but this
actually broke Razor rename in VS, because it runs in OOP, and the Razor
mapping service is not part of the composition. The fix is easy - Razor
will do the mapping when it gets the results - just need to make Roslyn
actually give the edits back to Razor so it can.

Diff best viewed with whitespace off.