Do not copy hoisted locals or parameters unnecessarily (#81227)
Instead of making a copy of an already hoisted by-local or parameter, we
should reuse the original local or parameter directly. This ensures that
we don't accidentally copy the original local or parameter
unintentionally and miss observing mutations to value types. Closes
https://github.com/dotnet/roslyn/issues/81212.

Relates to test plan https://github.com/dotnet/roslyn/issues/75960