Generate docs - fix an issue where the suggestion session was being overtaken (#81537)
The issue was that generate documentation was only intermittently
starting a Suggestion Session.
This was tracked down to inline completions taking control of the
Suggestion Session since it creates the session before making calls into
the LLM, while our generate documentation was waiting until after.

The fix:
Had to create the Suggestion Session early and prior to calling
`nextHandler()` so that we keep control. This also means we have to
check earlier whether or not the trigger character is at the last slash
to insert the doc comment.