Add willRenameFiles support, and an abstraction for extenders to implement (#81549)
Part of https://github.com/dotnet/razor/issues/8541

Adding functionality to Razor so that when a .razor file is renamed, the
relevant component tags and any C# references are updated, so need
willRename support. Since its a workspace request, if we implement it
then Roslyn won't be able to, so best to support it in Roslyn and
delegate to Razor as required.