SyntheticBoundNodeFactory: Remove ability to create arbitrary conversions  (#81569)
Language rules around conversions are getting more and more complex. At
the same time Convert APIs in SyntheticBoundNodeFactory might confuse
consumer into thinking that they are able to synthesize nodes for
arbitrary conversions, which is not the case.

This change removes one API and adds an assert to remaining API to
verify that only conversions that are natively supported by Emit layer
are coming through.