Ensure binder with `BinderFlags.AttributeArgument` is in the local binder chain (#81628)
Fixes #81576.

A binder with the flag was injected manually as the current binder while
binding attributes. Because of that `BinderFlags.AttributeArgument` was
not preserved by `GetBinder` API. For example the API is used during
binding of `nameof`. Various checks responsible for breaking infinite
binding cycles are relying on presence of the flag.

Since after the change the flag is preserved when we are binding
`nameof`, I had to add a fixup in several places to properly support
`IDS_FeatureInstanceMemberInNameof` feature in context of an attribute.
See #67461 that initially added support for the feature.