Give the unified EA a separate namespace.
For the time that both the old EA assemblies and the new unified assemblies will be loaded into the MEF composition, we need to use separate namespaces so that we do not have mulitple exports for the same type. The namespaces will need to be updated in our partners repos when they move over to the unified EA packages.