Clean up our LoadComponentsInUIContext helpers
We had some code scheduling to the UI thread to wait for the UI context
to be complete, but that's no longer necessary. Since we also had
multiple derived types all setting this up, but all of them had the
pattern in one way or another, this also just unifies the pattern
to just one case.