Improve error recovery around 'scoped' modifier parsing (#81636)
We now allow a misplacced 'scoped' keyword to be treated as a normal
modifier, and move teh reporting of issues to the binding phase. Today
we get entirely thrown off and enter very bad error recovery that often
gets thrown entirely off, creating tons of cascading errors.

Note: the net effect of this is much fewer errors, and much better error
recovery. The added length in this PR is the addition of new tests, or
updating existing tests to report binding errors to demonstrate that
errors are still produced semantically where they would have previously
been syntactic.